Move checkAndOverrideIsolationOptions to run_common.go
